Modify introduction in Slack article
The Slack guide was copied from the old Slack document.
Its introduction read:

```
Hi and welcome to Slack!

It's nice to have you here!
```

Such introduction would be fine for new colleagues who received
the link in Slack. For users outside the company however this would
make no sense - they would land on a web page, not in a Slack channel.

The introduction was modified to accomodate broader audience.